+++ app/src/main/java/com/example/orderdeliver/domain/exceptions/WrongPriceException.kt
@@ -1,3 +1,3 @@
 package com.example.orderdeliver.domain.exceptions
 
-class WrongPriceException(value: Int) : Exception("The price must be more than 0, but your price is $value")
+class WrongPriceException(value: Float) : Exception("The price must be more than 0, but your price is $value")
